// src/api/generated/models/publicPropertyControllerFindById200PhotosItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

export type PublicPropertyControllerFindById200PhotosItem = {
  url?: string;
  /** @nullable */
  caption?: string | null;
  cover?: boolean;
  /** @nullable */
  displayOrder?: number | null;
};